Updated documentation of VNFPackage
Since the image uploading of an image inside the VNFPackage is not working at the moment
it was decided to disable this feature temporarly. Also the VNFPackageManagement will be updated to reject/ignore
the uploading of an image from inside the Package. So at the moment the only way
to upload an image with the VNFPackage is to use the image link inside the Metadata.yaml file.